requirements.txt
# Core dependencies
aiogram==3.7.0
asyncpg==0.30.0
aiohttp==3.9.5
python-dotenv==1.0.1
pydantic-settings==2.3.4

# Database
sqlalchemy[asyncio]==2.0.35
alembic==1.13.1
greenlet==3.0.3
psycopg2-binary==2.9.9  # For Alembic migrations

# Time zones and scheduling
pytz==2024.1

# Logging
loguru==0.7.2

# Optional dependencies
redis==5.0.4
asyncio-mqtt==0.16.1

# Webhook server
aiohttp-cors==0.7.0

# Google Sheets integration
gspread==5.12.4
google-auth==2.29.0
